+++ packages/twenty-front/src/modules/page-layout/widgets/graph/graphWidgetGaugeChart/types/GaugeChartData.ts
@@ -1,0 +1,10 @@
+import { type GraphColor } from '@/page-layout/widgets/graph/types/GraphColor';
+
+export type GaugeChartData = {
+  value: number;
+  min: number;
+  max: number;
+  color?: GraphColor;
+  to?: string;
+  label?: string;
+};
