+++ packages/twenty-front/src/modules/page-layout/widgets/graph/graphWidgetNumberChart/types/NumberChartData.ts
@@ -1,0 +1,4 @@
+export type NumberChartData = {
+  value: string;
+  trendPercentage: number;
+};
