+++ packages/twenty-front/src/modules/page-layout/widgets/graph/graphWidgetNumberChart/utils/formatNumberChartTrend.ts
@@ -1,0 +1,3 @@
+export const formatNumberChartTrend = (trendPercentage: number): string => {
+  return trendPercentage >= 0 ? `+${trendPercentage}` : `${trendPercentage}`;
+};
