+++ packages/twenty-front/src/modules/page-layout/widgets/graph/graphWidgetPieChart/types/PieChartDataItem.ts
@@ -1,0 +1,9 @@
+import { type GraphColor } from '@/page-layout/widgets/graph/types/GraphColor';
+
+export type PieChartDataItem = {
+  id: string;
+  value: number;
+  label?: string;
+  color?: GraphColor;
+  to?: string;
+};
